reorganize test structure
